fix(harness): mask GIE during polled marker print
The mdb harness transmits the PASS/FAIL marker with a 6-deep polled
call chain (report -> log -> putc -> Transmit -> ClearFlag). The 1 ms
tick ISR adds 4 more levels (vector -> dispatch -> handler ->
callback), overflowing the PIC16 8-level hardware stack when it fires
mid-print. The corrupted return address makes the next stringdir fetch
return garbage, so the marker truncates at 21 bytes and the gate never
sees PASS.

Mask GIE for the whole log line in both PIC16 family harnesses. The
PIC18 harness is unaffected (31-level stack); the 193X harness drives
GPIO, not UART.

+++ pic16f88x-hal/src/mdb/pic16_harness_mdb.c
@@ -6,7 +6,7 @@
 
 #include "core/epic_harness.h"
 #include "core/pic16_irq.h"
-#include "peripherals/pic16f87xa_usart.h"
+#include "peripherals/pic16f88x_usart.h"
 
 #include <stdint.h>
 
@@ -23,7 +23,7 @@ static uint32_t g_cycles = 0U;
  */
 static void s_tx_cplt(void)
 {
-    /* Non-null so EPIC_USART_Init arms TXEN (pic16f87xa_usart.c); never
+    /* Non-null so EPIC_USART_Init arms TXEN (pic16f88x_usart.c); never
      * actually called, transmission below is polled. */
 }
 
@@ -56,7 +56,7 @@ void epic_harness_init(uint32_t cycles)
     g_cycles = cycles;
 
     s_usart_handle = (USART_HandleTypeDef)USART_HANDLE_DEFAULT;
-    s_usart_handle.SPBRG = (uint8_t)USART_ComputeSPBRG(
+    s_usart_handle.SPBRG = (uint16_t)USART_ComputeSPBRG(
         FOSC_HZ, EPIC_HARNESS_SIM_BAUD, USART_MODE_ASYNCHRONOUS,
         USART_BRGH_HIGH);
     s_usart_handle.TxCpltCallback = s_tx_cplt;
